pyproject.toml
[project]
name = "lang-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "skyrimer",email = "chekmenev2004@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]
